todo 작성중

+++ src/views/Playground.vue
@@ -1,34 +1,58 @@
 <template>
-  <div class="d-flex">
-    <e-input
-      v-model="value"
-      class="mr-2"
-      @keydown.enter="addTodo"
-    />
-    <e-button
-      @click="addTodo"
-    >
-      Enter
-    </e-button>
-  </div>
-  <div>
-    <template
-      v-for="(todo, index) in todos"
-      :key="index"
-    >
-      <div class="pt-2 pb-2">
-        {{ todo.text }}
-        <e-button @click="deleteTodo(todo)">
-          Delete
-        </e-button>
-      </div>
-    </template>
-  </div>
+  <section>
+    <div class="d-flex">
+      <e-input
+        v-model="value"
+        class="mr-2"
+        @keydown.enter="addTodo"
+      />
+      <e-button
+        @click="addTodo"
+      >
+        Enter
+      </e-button>
+    </div>
+    <div class="text-right mt-2">
+      <span class="mr-2">
+        Done: {{ progressCount }}
+      </span>
+      <span class="mr-2">
+        Left: {{ allCount - progressCount }}
+      </span>
+    </div>
+    <div>
+      <template
+        v-for="(todo, index) in todos"
+        :key="index"
+      >
+        <div
+          class="pt-2 pb-2 d-flex justify-space-between"
+        >
+          <div :class="[todo.done ? 'checked' : '']">
+            {{ todo.text }}
+          </div>
+          <div>
+            <e-button
+              class="mr-1"
+              @click="toggleTodo(todo)"
+            >
+              {{ todo.done ? 'Done' : 'Check' }}
+            </e-button>
+            <e-button
+              @click="deleteTodo(todo)"
+            >
+              Delete
+            </e-button>
+          </div>
+        </div>
+      </template>
+    </div>
+  </section>
 </template>
 
 <script lang="ts">
 
-import { ref } from 'vue'
+import { ref, computed, watch } from 'vue'
 import { useStore } from 'vuex'
 import EButton from '@/components/atoms/EButton.vue'
 import EInput from '@/components/atoms/EInput.vue'
@@ -45,29 +69,67 @@ export default {
     const todos: Todo[] = store.getters['todo/todos']
 
     const addTodo = () => {
+      const text = value.value
+      const id = generateId()
+      const done = false
+
+      if (text.length === 0) {
+        return window.alert('Please enter todo.')
+      }
+
       const todo: Todo = {
-        text: value.value,
-        id: generateId(),
-        complete: false
+        text,
+        id,
+        done
       }
 
       store.dispatch('todo/addTodo',todo)
+      value.value = ''
     }
 
     const deleteTodo = (todo: Todo) => {
+      const yes = window.confirm('Would you wanna remove this todo?')
+
+      if (!yes) {
+        return
+      }
+
       store.dispatch('todo/removeTodo', todo)
     }
+
+    const toggleTodo = (todo: Todo) => {
+      store.dispatch('todo/toggleTodo', todo)
+    }
+
+    const progressCount = computed(() => todos.filter((todo: Todo) => todo.done).length)
+    const allCount = computed(() => todos.length)
+
+    // watch(progressCount, () => {
+    //   if (progressCount.value === allCount.value) {
+    //     window.alert('You done all todo!')
+    //   }
+    // })
 
     return {
       value,
       todos,
+      allCount,
+      progressCount,
       addTodo,
-      deleteTodo
+      deleteTodo,
+      toggleTodo
     }
   }
 }
 </script>
 
 <style scoped lang="scss">
+  section {
+    max-width: 800px;
+    margin: auto;
+  }
 
+  .checked {
+    text-decoration: line-through;
+  }
 </style>
